pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dental-ai-treatment-planner"
version = "1.0.0"
description = "AI-Powered Dental Treatment Planner with YOLOv11, MCP Server, and RAG"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Kannan", email = "kannan@ibritz.co.uk" }
]
keywords = [
    "dental-ai",
    "yolo",
    "mcp",
    "rag",
    "healthcare",
    "computer-vision",
    "treatment-planning"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    # Core AI/ML
    "ultralytics>=8.3.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    
    # MCP Server
    "mcp>=1.0.0",
    "fastmcp>=2.0.0",
    
    # RAG Components
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    
    # API & Web
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.12",
    "gradio>=5.0.0",
    
    # Data Processing
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    
    # Utilities
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.scripts]
dental-ai = "src.cli:app"
dental-mcp = "src.mcp_server.server:main"

[project.urls]
Homepage = "https://github.com/ibritz/dental-ai-treatment-planner"
Documentation = "https://github.com/ibritz/dental-ai-treatment-planner#readme"
Repository = "https://github.com/ibritz/dental-ai-treatment-planner"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
